Moved the package from Editor folder to Scripts folder.
This process included the move of number sprites from an Editor subfolder to the Resources folder, since it is the best solution for loading images both from the editor and in the game runtime.
To make the code visible to tests, I also created two assembly definitions and used the internal VisibleTo attribute to make internal classes accessible to tests.